Reimplement simple refactoring encoding
Where there is one variable per node that could be in the refactoring.

Since the partition scheme of bicliques does not make sense with this
encoding, I implemented a partition based on index of the refactoring.
The ith partition of n partitions insists idx % n = i.

Can be enabled with --no-bicliques.