Fix emptiness bug with tracking of last position
E.g. :last-child ~ *

Because * had no position info recorded by the emptiness alg, because it
didn't need any at the point it was first seen (working backwards).

This meant the position of :last-child wasn't compared against anything.
So it wasn't spotted that it wasn't the last child.

Add a check in this case against the nlast position variable when there
are no tracked positions.